Move `BoundaryOnCurve` to `geometry`
I'd like to use it outside of the approximation code.
