+++ crates/fj-core/src/geometry/mod.rs
@@ -1,9 +1,11 @@
 //! Types that are tied to objects, but aren't objects themselves
 
+mod boundary;
 mod path;
 mod surface;
 
 pub use self::{
+    boundary::BoundaryOnCurve,
     path::{GlobalPath, SurfacePath},
     surface::SurfaceGeometry,
 };
